requirements.txt
cryptography
arabic-reshaper
python-bidi

